bank: transaction history
Added a transaction history functionality, so that users could see all of the transactions they have made. This will make it more convinient for them. Also it will help us with the statistics of the bank activity issues, because we will be able to get the actual data and display it. Button to the account page of the user will be added later.

Fixes #76